scripts/requirements.txt
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
beautifulsoup4>=4.12.0
azure-search-documents>=11.6.0
azure-identity>=1.17.0
openai>=1.40.0
